server/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Google Cloud dependencies
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Vertex AI
google-cloud-aiplatform==1.38.1

# LangChain dependencies
langchain==0.0.340
langchain-google-genai==0.0.6
langchain-community==0.0.2
langgraph==0.6.2
tavily-python==0.3.0

# Additional dependencies for your existing code
pinecone-client==2.2.4